requirements.txt
# Web (PWA)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.12

# Scheduler
apscheduler>=3.10.0

# LLM
anthropic>=0.40.0

# Ad Platforms
facebook-business>=20.0.0
google-ads>=28.0.0

# HTTP / Data
httpx>=0.27.0
feedparser>=6.0.0

# Config / Utils
python-dotenv>=1.0.0
rich>=13.0.0

# DB
pymongo>=4.9.0

# Auth
google-auth>=2.36.0
google-auth-oauthlib>=1.3.0
